+++ Nos/Extensions/Localization+Nos.swift
@@ -32,7 +32,8 @@ extension Localizable {
         }
     }
 
-    func localizedText(_ arguments: [String: String]) -> LocalizedStringKey {
+    /// Use this function in place of `text(_ arguments:[String: String]) -> String` to use Markdown-formatted text
+    func localizedMarkdown(_ arguments: [String: String]) -> LocalizedStringKey {
         LocalizedStringKey(text(arguments))
     }
